requirements.txt
python-jobspy>=1.1.0
google-genai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
beautifulsoup4>=4.12.0
